+++ src/main/java/org/example/chocostyle_datn/repository/ChatLieuRepository.java
@@ -3,5 +3,5 @@ package org.example.chocostyle_datn.repository;
 import org.example.chocostyle_datn.entity.ChatLieu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChatLieuRepository extends JpaRepository<ChatLieu, Long> {
+public interface ChatLieuRepository extends JpaRepository<ChatLieu, Integer> {
 }
