+++ src/main/java/org/example/chocostyle_datn/repository/DotGiamGiaRepository.java
@@ -3,5 +3,5 @@ package org.example.chocostyle_datn.repository;
 import org.example.chocostyle_datn.entity.DotGiamGia;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DotGiamGiaRepository extends JpaRepository<DotGiamGia,Long> {
+public interface DotGiamGiaRepository extends JpaRepository<DotGiamGia,Integer> {
 }
