+++ src/main/java/org/example/chocostyle_datn/repository/HoaDonRepository.java
@@ -3,5 +3,5 @@ package org.example.chocostyle_datn.repository;
 import org.example.chocostyle_datn.entity.HoaDon;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HoaDonRepository extends JpaRepository<HoaDon,Long> {
+public interface HoaDonRepository extends JpaRepository<HoaDon,Integer> {
 }
