+++ src/main/java/org/example/chocostyle_datn/repository/LichSuThanhToanRepository.java
@@ -3,5 +3,5 @@ package org.example.chocostyle_datn.repository;
 import org.example.chocostyle_datn.entity.LichSuThanhToan;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LichSuThanhToanRepository extends JpaRepository<LichSuThanhToan,Long> {
+public interface LichSuThanhToanRepository extends JpaRepository<LichSuThanhToan,Integer> {
 }
