+++ src/main/java/org/example/chocostyle_datn/repository/PhieuGiamGiaKhachHangRepository.java
@@ -3,5 +3,5 @@ package org.example.chocostyle_datn.repository;
 import org.example.chocostyle_datn.entity.PhieuGiamGiaKhachHang;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PhieuGiamGiaKhachHangRepository extends JpaRepository<PhieuGiamGiaKhachHang,Long> {
+public interface PhieuGiamGiaKhachHangRepository extends JpaRepository<PhieuGiamGiaKhachHang,Integer> {
 }
