+++ src/main/java/org/example/chocostyle_datn/repository/PhieuGiamGiaRepository.java
@@ -3,5 +3,5 @@ package org.example.chocostyle_datn.repository;
 import org.example.chocostyle_datn.entity.PhieuGiamGia;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PhieuGiamGiaRepository extends JpaRepository<PhieuGiamGia,Long> {
+public interface PhieuGiamGiaRepository extends JpaRepository<PhieuGiamGia,Integer> {
 }
