+++ src/main/java/org/example/chocostyle_datn/repository/SanPhamRepository.java
@@ -3,5 +3,5 @@ package org.example.chocostyle_datn.repository;
 import org.example.chocostyle_datn.entity.SanPham;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SanPhamRepository extends JpaRepository<SanPham,Long> {
+public interface SanPhamRepository extends JpaRepository<SanPham,Integer> {
 }
