+++ src/main/java/org/example/chocostyle_datn/repository/ThanhToanRepository.java
@@ -3,5 +3,5 @@ package org.example.chocostyle_datn.repository;
 import org.example.chocostyle_datn.entity.ThanhToan;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ThanhToanRepository extends JpaRepository<ThanhToan,Long> {
+public interface ThanhToanRepository extends JpaRepository<ThanhToan,Integer> {
 }
