+++ src/main/java/org/example/chocostyle_datn/repository/XuatXuRepository.java
@@ -3,5 +3,5 @@ package org.example.chocostyle_datn.repository;
 import org.example.chocostyle_datn.entity.XuatXu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface XuatXuRepository extends JpaRepository<XuatXu,Long> {
+public interface XuatXuRepository extends JpaRepository<XuatXu,Integer> {
 }
